requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
annotated-types==0.7.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
exceptiongroup==1.2.2
Faker==30.0.0
idna==3.10
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
packaging==24.1
pluggy==1.5.0
pydantic==2.9.2
pydantic_core==2.23.4
pytest==8.3.3
python-dateutil==2.9.0.post0
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
six==1.16.0
tomli==2.0.1
typing_extensions==4.12.2
urllib3==2.2.3
